Fixed errors and refactoring
